(tests): Move tests using `AWSHTTPConnection` to `test_aws_http_connection`
